Moved all non-file related logic from CompileShader.
CompilationContext::CompileShader contained a lot of
file specific logic as well as the compilation logic. This moves all of
the compilation logic into DoCompilation, and keeps all of the
file-related logic in CompileShader.

This is in preparation for moving all compilation logic into
shaderc_util to be shared with libshaderc.