Add shaderc_compile_options_set_target_env API
This provides an API for specifying the target shader environment
which should be used for warning/error generation. The default remains
Vulkan, but this enables use of legacy or non-Vulkan shaders.

The version parameter is currently unused, but is intended to distinguish
between different revisions of the target shader environment in the future.

This also adds SetTargetEnvironment() to C++ interface to expose the same
functionality.