Add compilation status to shaderc_spv_module
The StageDeducer in libshaderc now keeps a flag, set to true when it is
called to deduce a shader stage but failed (the last chance to get the
stage). We then check the flag after compilation, to determine the error
type, and store it in the result module

Replace get_success() with get_compilation_status(). Instead of
returning a boolean, it returns an enum type:
shaderc_compilation_status. The enum value can indicate whether the
compilation succeeded or failed due to some reasons.