When copying tests, only catch ENOENT exception
Before, we were catching too-broad a set of exceptions.
On Windows sometimes you might not be able to remove the
original directory, e.g. since a process might be running
in that directory.  That can lead to spurious test failures where
tests would run but compare to stale base results.

Also:
- sort the test file list, for nicer debugging.
- fix linting warnings for unused variables.