Updated CMake configuration for Android Studio.
Small change to fix how we are configured if included by Android Studio.
The main difference between Android Studio CMake, and
android-cmake, at least as far as this is concerned is that Android
Studio does not define find_host*. In this case we can just use
find_package and find_program, so set up the macros in that way.