Don't suppress stderr of subprocesses
When a subprocess failed in update_shaderc_sources.py, it was impossible
to know the failure's reason without hacking the script, because the
script suppressed the stderr of all subprocesses.

For example, when `git fetch` failed on the shaderc repo, the script
printed only this error message:

    Failed to run ['git', 'fetch', 'known-good'] in .

Is there no network connection? Did ssh authentication fail? Is the
repository corrupt? Who knows what happened?!?!?! Answer: stderr knows
what happened.