Re-add the 'known-good' remote if missing
If a subrepo exists but does not have a remote named 'known-good', then
update_shaderc_sources.py fails. This situation can happen if you try to
re-use an existing repo that you cloned manually without
update_shaderc_sources.py; or if you accidentally delete the
'known-good' remote.

To fix this, teach the script to check if the remote exists before
running git-fetch. If the remote is missing, then run `git remote add`.