glsl: Don't write dependency info when compilation output is stdout
If glslc is given option '-o -' and one of the -M* dependency
options, then it wrote an incorrect dependency file. For example:

    $ glslc -MD -S -o - a.glsl | sed ...
    $ ls
    -.d a.glsl
    $ cat ./-.d
    -: a.glsl

It doesn't make sense to write a dependency file when the compilation
output is stdout. So catch the error while validating command line
flags, emit a helpful error message, and fail.

Fixes: https://github.com/google/shaderc/issues/327