Introduce new option to skip installation
Define an option SHADERC_SKIP_INSTALL which depending on the value sets
SHADERC_ENABLE_INSTALL which will be used to test to bypass install()
rules in later projects.
Since this is a static library, I really only need to link to the libraries,
and not install them.

Chain skip install to third party libraries.

Also use GNUInstallDirs to install to lib64 where that's appropriate.