BUILD.gn: Make libraries Chromium-style "components"
Components are a concept defined in Chromium's BUILDCONFIG.gn that are
shared libraries (with the right rpath set up) or static_libraries
depending on whether is_component_build is defined.

This changes libshaderc and libshaderc_spvc to be components mostly so
that the right rpath options are set up automatically when they are
shared libraries.

Also fixes missing shaderc_util_public in the public_configs of both
libraries.