Avoid leading dot slash in cmake target.
Including shaderc in project that declares multiple build configurations
and then using generator that doesn't support multi-configuration would
result in a target path with a leading dot slash.  Ninja doesn't work
in that case.

Prepend CMAKE_CURRENT_BINARY_DIR in order to get an absolute path and
make ninja happy.  This works in all cases so we remove code specific
to multi-configuration.

Fixes #568.