Upate python scripts work with python3. (#569)
Make changes to the shaderc python scripts so they work with python2 and python3: fixes #474.

The Windows bots will use python3 so it is tests, while other bots will continue to use python2.

Changes were made to spirv-tools and spirv-cross to work with python3.  The DEPS file was updated to include these changes.

When using python2, the future package is now required, and that is reflected in the README.

There are some problems with spriv-cross.  The lastest version deleted the andriod.mk file, so we cannot build spvc in the NDK build anymore, so that was turned off.  Spirv-cross is using `nl_langinfo`, but this is in the andriod NDK starting with API version 26 (Android 8 I believe).  Since we do not have precedent for picking a specific version of android to build for, I disabled the build of spvc for the android bots.  This can be enabled again if we decide that the rest of shaderc does not need to be built for older versions.

Also, we do not plan on shipping spirv-cross in the android ndk, so that played a role in the decision.

Finally, the spvc tests were disabled.  They are based on the tests in spirv-cross, which are specific to the version of spirv-tools and glslang.  Many of them broke when updating spirv-cross.  They are disabled for now and @fjhenigman will look into fixing them.