Add -fnan-clamp
Tells the compiler to generate code for max and min builtins so that
when given a NaN operand, the builtin returns the other operand.
Similarly, the clamp builtin will favour the non-NaN operands, as
if clamp were implemented as a composition of max and min.

Requires Glslang commit a9b00ac5d58878c6e67cb50c2af93c170648d6c9