Extract common test enviroment bits out into a class
Refactors all of the various bits of common state and operations for
the tests into a single class. This reduces the amount of arg passing
that needs to occur, which in turn makes the code more readable.

The arguments to the script are filtered through this new class, so
that test code does not need to directly touch the parsed args, which
can be error prone.

Additionally the inputs to the script are changed from being all of
the binaries to use + the test file directory, to the directory where
the binaries are built + the test file directory.

Finally, the existing logging path is commented out, since it is
mostly broken under multiprocessing. An issue for fixing this has been
filed.

Fixes #735