Catch low level exceptions instead of crash
This patch refactors calls to system libraries that might throw to
catch and return a failure instead of crashing.

This make debugging easier when tests fail due to invalid spirv, etc
since now an unexpected failure is reported instead of a stack trace.

This also lays some of the ground work for refactoring logging, but
having a path for command output to travel through.

Catching is performed at the callsite, instead at the top of the work
function, so that catching ca be specific to the exceptions that are
used to indicate expected failures. Other exceptions in this script
suggest a bug or programming error.

Fixes #740