Extract common test environment bits out into a class (#738)
Refactors all of the various bits of common state and operations for
the tests into a single class. This reduces the amount of arg passing
that needs to occur, which in turn makes the code more readable.

The arguments to the script are filtered through this new class, so
that test code does not need to directly touch the parsed args, which
can be error prone.

Finally, the existing logging path is commented out, since it is
mostly broken under multiprocessing. An issue for fixing this has been
filed.

Fixes #735