Fix some Python2 vs 3 issues in update_build_version.py (#783)
Cleans up a usage of print.

Also changes the behaviour of reading in the CHANGES file to insert
'?' when encountering non-ASCII, instead of throwing. Given that this
part of the code is just extracting the version string from CHANGES,
this should not have unicode in it. If a ? is getting inserted in
there, then the file is malformed and this operation will fail.

Fixes #778