The glslang-testsuite is moved into glslang. (#838)
Currently ShaderC does a full copy of the glslang testsuite during the build, rewrites the test script and then runs the tests as part of ctest. The code to run those tests has been added to the GLSLang CMake script so will be added to ctest automatically.

This CL removes the glslang-testsuite from shaderc and bumps the version of glslang so we will now use the version provided directly.