Split spvc shader generation into initialize and compile phases
The current API does all of the validation, IR generation, and cross
compile in a single operation. Third party users, specifically Dawn,
want to be able to insert their own commands between the IR generation
and compilation stages. Thus the API is being broken up into two
commands, an initialization command that validates input and generates
the IR, and a compile command that actually performs the cross
compile.

Fixes #863

Trying to fix the test