Another major API rework (#871)
Move messages string out of the compilation result struct into the
context struct.

Also move the status entry out of the result struct. Instead have
functions that were returning the result struct, take in a result
struct & return a status. This significantly changes the lifetime of
the result to be similar to other opaque structs. Now the caller is
responsible for creating the result.

Fixes #867