Convert compile options to require explicit environments
This is designed to replace the existing code which has default values, and
setters, with a pattern that requires specifying the environments at creation
time.

All of the internal code and tests has been updated to use the new API.

The old API has been marked as a deprecated and will be removed once Dawn no
longer uses it.

Part of #1017