Standardised git-sync-deps.py
The edited command before changes is incompatible with Windows. There is no file extension to identify the type of script it is running and no explicit invocation with python. There is no path and version number to be specified at the top of the file or in the command because this can differ between machines. It either works or it doesn't. If it doesn't, the machine is not tolerant to new standards and needs an update or the repository repackaging.