build: add switch to disable copyright check
A Vulkan SDK build of shaderc fails on the copyright check:

    C:\Program Files (x86)\MSBuild\Microsoft.Cpp\v4.0\V140\Microsoft.
    CppCommon.targets(171,5): error MSB6006: "cmd.exe" exited with co
    de 1. [C:\j\msdk\build\shaderc\repo\src\dbuild32\check-copyright.
    vcxproj]

Analysis in Visual Studio reveals that the problem is:

    20>  .\dbuild\include\glslang\build_info.h  has no copyright mess
    age.
    20>C:\Program Files (x86)\MSBuild\Microsoft.Cpp\v4.0\V140\Microso
    ft.CppCommon.targets(171,5): error MSB6006: "cmd.exe" exited with
    code 1.

"build_info.h" does indeed have a copyright message:

    // Copyright (C) 2020 The Khronos Group Inc.

but this fails the copyright check, which is looking solely for:

    Copyright <NNNN> The Shaderc Authors. All rights reserved.

These changes add a SHADERC_SKIP_COPYRIGHT_CHECK option that can be
passed at generation time to disable the copyright check and allows
the build to succeed.

It turns out CMake 3.10 ordered the shaderc build in such a way that
the copyright check occurred at the end of the build, while CMake 3.19
puts the copyright check at the beginning of the build.  So the SDK
*could* be built with CMake 3.10 if the build error was ignored, but
cannot be built with CMake 3.19 even if the build error is ignored.