Don't use generator expressions
While generator expression is more concise, it is avoidable in this
case. This change fixes compatibility with Meson, which currently does
not support generator expressions and would incorrectly enable install.
While this shouldn't be critical, it exposes other compatibility issues.
In short, by changing this basic expression, I'm able to build shaderc
with Meson wrap successfully.